reftek: move new implementation in place of initial one + refactoring + cleanup
